Use new cache accessors to increase performances.
A common pattern is to write on a file one byte at a time.
Since consecutive writes are mostly on the same block, BattFS
would allocate a new block each time to avoid burning
the sector.
This is very inefficient, and since we have a cache, we can
avoid re-writing the block by using it.

This patch does not rewrite a block if it is cached, and use
the cache efficiently.



git-svn-id: https://src.develer.com/svnoss/bertos/trunk@3994 38d2e660-2303-0410-9eaa-f027e97ec537