Flush the transfer before exiting from write function.
Since the buffer is used by the DMA, if we return early we run the
risk of having the buffer overwritten before the transfer will be
completed.
Waiting for the flush could be theorically a pessimization, but from
tests, since the typical use involves mixed reads/writes, we already
wait for the transfers to be completed, and performance are
almost the same.


git-svn-id: https://src.develer.com/svnoss/bertos/trunk@4017 38d2e660-2303-0410-9eaa-f027e97ec537