mware: fix generic completion events behaviour without the kernel
The correct behaviour of the generic completion events is to allow any
process to wait for the event and not only the one that initialized the
completion.

With this fix the "owner" of the completion is assigned when a process
call event_wait() and not event_initGeneric(). In this way a call to
event_do() correctly wakes up the process that actually performed the
event_wait().

Moreover, use the signal SIG_SYSTEM5 for generic completion events
(instead of SIG_SINGLE, that is already by the timer and msg).

git-svn-id: https://src.develer.com/svnoss/bertos/trunk@4463 38d2e660-2303-0410-9eaa-f027e97ec537