Remove kludge for c++.
The kludge removed warnings when compiling files with c++,
used only in the demo.
However, if the user uses a path to the compiler which
contains '++' somewhere (eg.
CodeSourcery/Sourcery_G++_Lite/bin/arm-none-eabi-"gcc"), the
C specific flags are removed.

This patch solves the issues by clearing out the C specific
flags only in the demo.


git-svn-id: https://src.develer.com/svnoss/bertos/trunk@4643 38d2e660-2303-0410-9eaa-f027e97ec537