STM32: USB: correctly detect when read/write must be called without blocking
Read or write operations with the EP0 may happen within the USB
interrupt.

Make the driver able to understand when the usb_endpointRead() or
usb_endpointWrite() is called from the interrupt context or not.

Basically, only the even callbacks concerning the EP0 should be called
in a interrupt context. For all the other cases the user must take care
of using usb_endpointRead() and usb_endointWrite() from a non-atomic
(sleepable) context.

git-svn-id: https://src.develer.com/svnoss/bertos/trunk@4717 38d2e660-2303-0410-9eaa-f027e97ec537