Add a priority inversion test for Semaphores.
Introduce a test case in sem_test.c for a typical priority
inversion situation where some high and low priority processes
share one or more semaphores and are disturbed by a non-interacting
medium priority process. The test shows, by logging the finishing
time of the procecsses, that the execution of an high priority
process can be noticeably delayed by someone with lower priority
because of interactions happening on semaphores.

Signed-off-by: Dario Faggioli <raistlin@linux.it>

git-svn-id: https://src.develer.com/svnoss/bertos/trunk@4907 38d2e660-2303-0410-9eaa-f027e97ec537